+++ src/CloudToppedMixedLayerModel/free_troposphere.jl
@@ -3,15 +3,16 @@
 
 Return an equation for ``T_{FTR}``. `γ` is as in [Datseris2025](@cite).
 `add_co2` will add an additional warming term `ECS_CO2*log2(CO2/400)`.
+Introduces two extra parameters, `T_FTR_0, T₊_0`.
 """
-function free_troposphere_emission_temperature(version = 1.0; add_co2 = true)
+function free_troposphere_emission_temperature(version = 1.0; add_co2 = true, T₊_0_init = 5.0)
     # We measure the subtropical component of T_FTR from T₊
     # not from T_t, because that's what makes sense due to physical continuity.
     # T₊ already has a CO2 component so we are careful here to add it only
     # to the tropical component of T_FTR!
     @parameters begin
         (T_FTR_0 = 290.0), [description = "prescribed temperature emission of free troposphere without CO2 effects, K"]
-        (T₊_0 = 5.0), [description = "Subtracted temperature from T₊ when defining T_FTR as a T₊, K"]
+        (T₊_0 = T₊_0_init), [description = "Subtracted temperature from T₊ when defining T_FTR as a T₊, K"]
     end
     if version == :strong # strong influence of subtropical
         weight = 0.0
@@ -41,12 +42,14 @@ end
         CO2_effect = false,
     )
 
-Provide equation for ``s_+``. To do this, a boundary condition must be provided
-that is a fixed parameter. `version` argument decides this:
+Provide equation for ``s_+`` depending on `version` with options:
 
 - `:difference`: the temperature difference across inversion is a fixed parameter.
 - `:temperature`: the temperature after the inversion is a fixed parameter.
 - `:static_energy`: the moist static energy after the inversion is a fixed parameter.
+- `:lapse_rate`: the temperature after the inversion is fixed and given by a prescribed
+  lapse rate, ``T_+ = T_{+,ref} + \\Gamma_T (z_b - 1000)``  as in [Salazar2023](@cite)
+  which introduces two additional parameters: `Γ_T = 6.5e-3, T₊_ref = 290.0`
 
 Besides these, we can also specify whether CO2 increase also increases temperature difference,
 and whether decreasing ``C`` decreases temperature difference due to cloud thinning
@@ -61,6 +64,8 @@ function mlm_s₊(
         (Δ₊T_C = 10.0), [description = "temperature decrease in the inversion due to cloud thinning (max lost K for 100% C.F.), K"]
         (s₊_0 = 300.0), [description = "prescribed moist static energy above inversion normalized by cₚ, K"]
         (T₊_0 = 292.0), [description = "prescribed temperature above inversion without CO2 or cloud effects, K"]
+        (T₊_ref = 290.0), [description = "reference temperature above inversion (for lapse rate), K"]
+        (Γ_T = 6.5e-3), [description = "free tropospheric temperature lapse rate, K/m"]
     end
 
     # First, prepare the augmentation of the inversion
@@ -85,13 +90,16 @@ function mlm_s₊(
         eqs = [
             T₊ ~ T₊_0 + Δ₊T_aux,
             s₊ ~ T₊ + g*z_b/cₚ,
-            Δ₊T ~ T₊ - T_t, # observable
         ]
-    elseif fixed == :static_energy
+    elseif inversion_fixing == :static_energy
         eqs = [
             s₊ ~ s₊_0 + Δ₊T_aux,
             T₊ ~ s₊ - g*z_b/cₚ,
-            Δ₊T ~ T₊ - T_t, # observable
+        ]
+    elseif inversion_fixing == :lapse_rate
+        eqs = [
+            T₊ ~ T₊_ref - Γ_T*(z_b - 1000.0),
+            s₊ ~ T₊ + g*z_b/cₚ,
         ]
     else
         error("incorrect specification for what to stay fixed!")
@@ -103,14 +111,22 @@ end
     mlm_q₊(version = :relative)
 
 Provide equation for ``q_+``. If `version = :relative` then
-make free tropospheric relative humidity a free parameter.
-Else if `version = :constant` then make ``q_+`` itself a parameter.
+make free tropospheric relative humidity `RH₊` a free parameter.
+Else if `version = :constant` then make `q₊` itself a parameter.
+Else if `version = :lapse_rate` prescribe ``q_+ = q_{+, ref} - \\Gamma_q(z_b - 1000)``
+which introduces parameters `Γ_q = 1.5e-3, q₊_ref = 2.0`.
 """
 function mlm_q₊(humidity_fixing = :relative)
     if humidity_fixing == :relative
         return q₊ ~ RH₊ * q_saturation(T₊)
     elseif humidity_fixing == :constant
         return ParameterProcess(q₊, 1.5)
+    elseif humidity_fixing == :lapse_rate
+        @parameters begin
+            (q₊_ref = 2.0), [description = "reference specific humidity above inversion (for lapse rate), q/kg"]
+            (Γ_q = 1.5e-3), [description = "specific humidity lapse rate, g/kg/m"]
+        end
+        return q₊ ~ q₊_ref - Γ_q * (z_b - 1000.0)
     end
 end