+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# v1.2
+
+## `CloudToppedMixedLayerModel`
+
+- Added lapse rate option for s+ and q+ when coupling to the free troposphere
+- Free temperature emission temperature T_FTR now defaults to T₊ if not given an equation.
+
 # v1.1
 
 ## `CloudToppedMixedLayerModel`